app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
anthropic==0.40.0
python-dotenv==1.0.0
mangum==0.19.0
